+++ packages/docs/src/components/DownloadWalletZip.tsx
@@ -2,5 +2,5 @@ import { WALLET_DOWNLOAD_PATH } from '../constants';
 import { Link } from './Link';
 
 export function DownloadWalletZip() {
-  return <Link href={WALLET_DOWNLOAD_PATH}>Bacon Wallet zip file</Link>;
+  return <Link href={WALLET_DOWNLOAD_PATH}>Jason Wallet zip file</Link>;
 }