+++ packages/docs/src/components/FuelBrandingDropdown.tsx
@@ -10,9 +10,9 @@ import {
 import { HeaderFuelBranding } from './HeaderFuelBranding';
 
 const environmentsTitles: Record<Environment, string> = {
-  [Environment.PRODUCTION]: 'Bacon Wallet',
-  [Environment.NEXT]: 'Bacon Wallet Next',
-  [Environment.STAGING]: 'Bacon Wallet Development',
+  [Environment.PRODUCTION]: 'Jason Wallet',
+  [Environment.NEXT]: 'Jason Wallet Next',
+  [Environment.STAGING]: 'Jason Wallet Development',
 };
 
 const environments: Array<Environment> = Object.values(Environment);