+++ packages/docs/src/components/HeaderFuelBranding.tsx
@@ -2,7 +2,7 @@ import { cssObj } from '@fuel-ui/css';
 import { Box, FuelLogo } from '@fuel-ui/react';
 
 export function HeaderFuelBranding({
-  title = 'Bacon Wallet',
+  title = 'Jason Wallet',
 }: { title: string }) {
   return (
     <Box.Flex css={{ alignItems: 'center' }}>