+++ packages/docs/src/hooks/useExtensionTitle.tsx
@@ -3,5 +3,5 @@ import { useIsPreviewEnv } from './useIsPreviewEnv';
 export function useExtensionTitle() {
   const isPreview = useIsPreviewEnv();
 
-  return isPreview ? 'Bacon Wallet Development' : 'Bacon Wallet';
+  return isPreview ? 'Jason Wallet Development' : 'Jason Wallet';
 }