+++ packages/connections/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Major Changes
 
-- [`c8344e3b`](https://github.com/FuelLabs/fuels-wallet/commit/c8344e3bf0ad98bbd1454a9b602445b72757178f) Thanks [@nelitow](https://github.com/nelitow)! - Bacon Wallet name change.
+- [`c8344e3b`](https://github.com/FuelLabs/fuels-wallet/commit/c8344e3bf0ad98bbd1454a9b602445b72757178f) Thanks [@nelitow](https://github.com/nelitow)! - Jason Wallet name change.
 
 ## 1.0.0